+++ src/algoritmos/ordenacao/InsertionSort.h
@@ -1,6 +1,6 @@
 #ifndef INSERTIONSORT_H
 #define INSERTIONSORT_H
 
-void insertionSort(int *vet, int size, int *numComp, int *numReg);
+void insertionSort(int *vet, int size, unsigned long long int *numComp, unsigned long long int *numReg);
 
 #endif // INSERTIONSORT_H
