+++ src/algoritmos/ordenacao/quickSortMediano.cpp
@@ -11,7 +11,7 @@ void gerarVetor(int* v, int k, int* aux){
 
 }
 
-int medianadek(int* vetor, int k, int* trocas, int *comparacoes){
+int medianadek(int* vetor, int k, unsigned long long int* trocas, unsigned long long int *comparacoes){
 
     quickSortRecursivoIds(vetor, 0, k-1, comparacoes, trocas);
 
@@ -21,7 +21,7 @@ int medianadek(int* vetor, int k, int* trocas, int *comparacoes){
 
 }
 
-void quickSortMediano(int *vetor, int inicio, int fim, int k, int *trocas, int *comparacoes) {
+void quickSortMediano(int *vetor, int inicio, int fim, int k, unsigned long long int *trocas, unsigned long long int *comparacoes) {
 
 
     if(inicio < fim){
